Add server for GCP deploy

+++ server/controllers/article.js
@@ -10,12 +10,12 @@ class Controller {
             .populate("author")
             .populate("comments")
             .then((data) => {
-                res.status(200).json({message: 'Data retreival sucess', data:data})
+                res.status(200).json({ message: 'Data retrieval success', data: data })
 
             })
             .catch((err) => {
                 console.log(err)
-                res.status(500).json({message: err, note: "See console for details"})
+                res.status(500).json({ message: err, note: "See console for details" })
             })
     }
     static createArticle(req, res) {
@@ -24,37 +24,42 @@ class Controller {
             .then((decoded) => {
                 //check if user exist in database
                 User
-                    .findOne({email: decoded.email})
+                    .findOne({ email: decoded.email })
                     .then((data) => {
-                        if(data) {
-                            return Article
-                            .create({
-                                name: req.body.name,
-                                body: req.body.body,
-                                author: req.body.authorId
-                            })
-                            .then((result) => {
-                                //push article to user "Articles" property
-                                let currentArticle = data.articles
-                                currentArticle.push(result)
-                                User
-                                    .updateOne({
-                                        email: data.email
-                                    },{
-                                        articles: currentArticle
-                                    })
-                                    .then((data) => {
-                                        res.status(201).json({message: "Article successfully created", data: result})
-                                    })
-                            })
+                        if (data) {
+                            //check if the decoded author is the same as the one in req.body
+                            if(decoded._id == req.body.authorId) {
+                                return Article
+                                .create({
+                                    name: req.body.name,
+                                    body: req.body.body,
+                                    author: req.body.authorId
+                                })
+                                .then((result) => {
+                                    //push article to user "Articles" property
+                                    let currentArticle = data.articles
+                                    currentArticle.push(result)
+                                    User
+                                        .updateOne({
+                                            email: data.email
+                                        }, {
+                                                articles: currentArticle
+                                            })
+                                        .then((data) => {
+                                            res.status(201).json({ message: "Article successfully created", data: result })
+                                        })
+                                })
+                            } else {
+                                res.status(403).json({ message: "You are not authorized to create this article" })
+                            }
                         } else {
-                            res.status(400).json({message: "Invalid user"})
+                            res.status(400).json({ message: "Invalid user" })
                         }
                     })
             })
             .catch((err) => {
                 console.log(err)
-                res.status(500).json({message: err.message})
+                res.status(500).json({ message: err.message })
             })
     }
     static getUserArticle(req, res) {
@@ -63,19 +68,19 @@ class Controller {
             .then((decoded) => {
                 //check if user exist in database
                 User
-                    .findOne({email: decoded.email})
+                    .findOne({ email: decoded.email })
                     .populate("articles")
-                    .then((data) => {       
-                        if(data) {
-                            res.status(200).json({message: "Data retrieval success", data: data.articles})
+                    .then((data) => {
+                        if (data) {
+                            res.status(200).json({ message: "Data retrieval success", data: data.articles })
                         } else {
-                            res.status(400).json({message: "Invalid user"})
+                            res.status(400).json({ message: "Invalid user" })
                         }
                     })
             })
             .catch((err) => {
                 console.log(err)
-                res.status(500).json({message: err.message})
+                res.status(500).json({ message: err.message })
             })
     }
     static editArticle(req, res) {
@@ -83,32 +88,39 @@ class Controller {
             .then((decoded) => {
                 //check if user exist in database
                 User
-                    .findOne({email: decoded.email})
-                    .then((data) => {       
-                        if(data) {
-                            return Article
+                    .findOne({ email: decoded.email })
+                    .then((data) => {
+                        if (data) {
+                            //check if the article that is going to be edited belong to the token user
+                            //find the article in the user data
+                            let indexofArticle = data.articles.indexOf(req.body.articleId)
+                            if (String(req.body.articleId) === String(data.articles[indexofArticle])) {
+                                return Article
                                     .findOneAndUpdate(
                                         {
-                                            _id:req.body.articleId
+                                            _id: req.body.articleId
                                         },
                                         {
-                                            $set:{
+                                            $set: {
                                                 name: req.body.name,
                                                 body: req.body.body,
                                                 updatedDate: new Date()
                                             }
                                         })
-                                        .then((result) => {
-                                            res.status(200).json({message: "Article successfully edited"})
-                                        })
+                                    .then((result) => {
+                                        res.status(200).json({ message: "Article successfully edited" })
+                                    })
+                            } else {
+                                res.status(403).json({ message: "You are not authorized to edit this comment" })
+                            }
                         } else {
-                            res.status(400).json({message: "Invalid user"})
+                            res.status(400).json({ message: "Invalid user" })
                         }
                     })
             })
             .catch((err) => {
                 console.log(err)
-                res.status(500).json({message: err.message})
+                res.status(500).json({ message: err.message })
             })
     }
     static deleteArticle(req, res) {
@@ -117,37 +129,44 @@ class Controller {
             .then((decoded) => {
                 //check if user exist in database
                 User
-                    .findOne({email: decoded.email})
+                    .findOne({ email: decoded.email })
                     .then((data) => {
-                        let indexofDelete = data.articles.indexOf(req.body.articleId)
-                        let newArray = data.articles.splice(indexofDelete,1)
-                        if(data) {
-                            return Article
-                                    .deleteOne({_id: req.body.articleId})
+                        //create new article array without the deleted article
+                        let newArray = data.articles.filter(article => article != req.body.articleId)
+                        if (data) {
+                            //to find the index of the article in "articles" propertie of the user
+                            let indexofArticle = data.articles.indexOf(req.body.articleId)
+                            // if the article we want to delete is in the "articles" list of the decoded user, then we delete the article and update the user's "articles"
+                            if (String(req.body.articleId) === String(data.articles[indexofArticle])) {
+                                return Article
+                                    .deleteOne({ _id: req.body.articleId })
                                     .then((result) => {
                                         //delete the article from the user "Article list"
-                                        return User 
-                                                .updateOne(
-                                                    {
-                                                        email: decoded.email
-                                                    },
-                                                    {
-                                                        articles: newArray
-                                                    })
-                                                    .then(() => {
-                                                        res.status(200).json({message: "Article successfully deleted"})
-                                                    })
+                                        return User
+                                            .updateOne(
+                                                {
+                                                    email: decoded.email
+                                                },
+                                                {
+                                                    articles: newArray
+                                                })
+                                            .then(() => {
+                                                res.status(200).json({ message: "Article successfully deleted" })
+                                            })
                                     })
+                            } else {
+                                res.status(403).json({ message: "You are not authorized to edit this comment" })
+                            }
+
                         } else {
-                            res.status(400).json({message: "Invalid user"})
+                            res.status(400).json({ message: "Invalid user" })
                         }
                     })
             })
             .catch((err) => {
                 console.log(err)
-                res.status(500).json({message: err.message})
+                res.status(500).json({ message: err.message })
             })
-
     }
 }
